Correction for all previous commit
There were some stupid mistak in previous commits and I modified it now
