+++ scribus/pdflib_core.cpp
@@ -62,7 +62,7 @@ for which a new license (GPL+exception) is in place.
 #include "commonstrings.h"
 #include "pageitem.h"
 #include "pageitem_textframe.h"
-#include "pageitem_textframe.cpp"
+//#include "pageitem_textframe.h"
 #include "pageitem_group.h"
 #include "pageitem_pathtext.h"
 #include "pageitem_table.h"
